+++ src/core/pages/users-profile-details/general.tsx
@@ -1,0 +1,11 @@
+import React from 'react'
+
+function General() {
+  return (
+    <div className='bg-red-500 '>
+      General
+    </div>
+  )
+}
+
+export default General
